drivers: thermal: Initialize thermal core earlier
CPUFreq drivers are available as part of subsys_initcall. In order to
facilitate thermal mitigation as early as possible in the boot cycle, move
the thermal framework initialization to subsys_initcall.

However, netlink initialization happens only as part of fs_initcall. At
this time in the boot process, the userspace would not be available though.
So initialize the netlink events later in fs_initcall.

Signed-off-by: Lina Iyer <ilina@codeaurora.org>
Signed-off-by: Amit Kucheria <amit.kucheria@linaro.org>