PM / devfreq: Actually support providing freq_table
The code in devfreq_add_device() handles the case where a freq_table is
passed by the client, but then frequests min and max frequences from
the, in this case absent, opp tables.

Read the min and max frequencies from the frequency table, which has
been built from the opp table if one exists, instead of querying the
opp table.

Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>