interconnect: Introduce per-provider locks
Currently, there is a single lock, the icc_lock, that used for handling all
icc_set_bw() requests, including those for unrelated paths. This means that
some request are unnecessarily serialized, which may in turn have impact on
performance.
To address this issue, we are introducing a provider specific locks.

Signed-off-by: Georgi Djakov <georgi.djakov@linaro.org>