media: venus: pm: Define v6 PM operations
This commit make a pm_ops structure for the V6 silicon. As an initial pass
it copies the v4 pm_ops structure and callbacks.

For the most part the logic for both IP blocks looks the same however there
are a few differences we will need to capture for v6. Existing v4
operations should cover most of it.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>