bus: mhi: Make mhi_state_str[] array static const and move to common.h
mhi_state_str[] array could be used by MHI endpoint stack also. So let's
make the array as "static const" and move it inside the "common.h" header
so that the endpoint stack could also make use of it. Otherwise, the
structure definition should be present in both host and endpoint stack and
that'll result in duplication.

Signed-off-by: Manivannan Sadhasivam <manivannan.sadhasivam@linaro.org>