PCI: qcom: Handle MSI IRQs properly
On Qualcomm platforms each group of MSI interrupts is routed to the
separate GIC interrupt. Thus to receive higher MSI vectors properly,
we have to setup and chain more MSI interrupts. However to remain
compatible with existing DTS files, do not fail if the platform doesn't
provide all 8 MSI interrupts. Instead of that, limit the amount of
supported MSI vectors.

Fixes: 8ae0117418f3 ("PCI: qcom: Add support for handling MSIs from 8 endpoints")
Signed-off-by: Dmitry Baryshkov <dmitry.baryshkov@linaro.org>