arm64: dts: Add msm8939 venus encoder/decoder hardware
The msm9836/msm8939 has two available encoder/decoder capable blocks. In
upstream we support a dedicated encoder and a dedicated decoder.

As far as I understand it, it would be feasible to have two decoders or two
encoders for parallel work like say decoding a video stream to each of the
DSI outputs.

Here though we add-in as per the upstream standard of fixed encoder and
fixed decoder at the first and second co-processor blocks.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>