dt-bindings: msm: dsi: Remove redundant phy-names
Adding in msm8939 which is based msm8916 dtsi I stumbled across a binding
check complaining about the phy name for msm8916 which we were reusing for
msm8939.

The currently inconsistent upstream dtsi naming of "dsi" and "dsi-phy" is
not captured in the yaml for this driver.

The driver however doesn't care about the name of DSI phy, hence the yaml
check is redundant.

Both Krzysztof and Rob suggested we could drop the phy-names entirely if it
really isn't a dependency.

So, drop the inconsistent and unnecessary phy-names field from the yaml and
dtsi for the qcom dsi phy. Do the drop for the yaml and dtsi in one patch
to avoid throwing up binding check errors with the robots.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>