media: venus: pm: Add V1 individual core get and power callbacks
msm8936/msm8939 have 2 x HFI 1xx cores which can be arranged as encoder and
decoder like versions 3xx and later.

Similar to 3xx and later we have
- A power domain per core
- A clock per core

But unlike 3xx we do not have a specific encoder start/stop command
address.

Add in callbacks for HFI V1xx compliant hardware to optionally set a
power-domain and clock for the encoder and decoder cores individually.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>